backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["cmrd-a <cmrd.a@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "2.2.10"
djangorestframework = "^3.11.0"
drf-yasg = "^1.17.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
